% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psjournals}
\alias{psjournals}
\title{A Dataset of Article Types in Political Science Journals}
\format{
A data frame with 717 rows and 16 variables:
\describe{
\item{\code{journal_id}}{integer, id number of the journal}
\item{\code{journal}}{character, name of the journal}
\item{\code{ssci_rank}}{integer, Web of Science's SSCI index (in the field of political science) of the journal, based on the 2-year Journal Impact Factor 2018}
\item{\code{h5_index}}{integer, Google Scholar's H5 Index of the journal, based on the articles published in 2014-2018}
\item{\code{h5_median}}{integer, Google Scholar's H5 Index of the journal, based on the articles published in 2014-2018}
\item{\code{since}}{integer, year of establishment of the journal}
\item{\code{publisher}}{character, publisher of the journal}
\item{\code{url}}{character, URL link to the journal website}
\item{\code{scope}}{character, scope of the journal}
\item{\code{type_id}}{integer, id number of the article type, by each journal}
\item{\code{type_category}}{character, category of the article type}
\item{\code{type}}{character, name of the article type, as defined by the journal}
\item{\code{lower_limit}}{character, lower (character/word/page) limit of the format}
\item{\code{upper_limit}}{character, upper (character/word/page) limit of the format}
\item{\code{limit_unit}}{character, unit of the type limits (character/word/page)}
\item{\code{last_updated}}{date, the date when each observation was last updated, in the "YYYY-MM-DD" format}
}
}
\source{
\code{ssci_rank} comes from \url{https://ooir.org},
\code{h5_index} and \code{h5_median} come from \url{https://scholar.google.ch/},
all remaining variables are collected from journal websites.
}
\usage{
psjournals
}
\description{
This is a dataset of article types published in selected political science journals.
}
\keyword{datasets}
